Move IsSpaceAvailable specializations to fwd/Kokkos_Fwd files
