+++ core/src/fwd/Kokkos_Fwd_SYCL.hpp
@@ -53,5 +53,13 @@ class SYCLSharedUSMSpace;  ///< Memory space accessible from both the SYCL
                            ///< device and the host
 class SYCL;                ///< Execution space for SYCL
 }  // namespace Experimental
+#if !defined(KOKKOS_ENABLE_SYCL)
+template <>
+struct IsSpaceAvailable<Experimental::SYCL> : std::false_type {};
+template <>
+struct IsSpaceAvailable<Experimental::SYCLDeviceUSMSpace> : std::false_type {};
+template <>
+struct IsSpaceAvailable<Experimental::SYCLSharedUSMSpace> : std::false_type {};
+#endif
 }  // namespace Kokkos
 #endif
